requirement.txt
pip install -U \
torch \
transformers \
sentence-transformers \
faiss-cpu \
pypdf
